requirements.txt
Django==4.1
asgiref==3.5.2
names==0.3.0
Pillow==9.0.1
dj-static==0.0.6
django-filter==21.1
django-utils-six==2.0
djangorestframework==3.13.1
pytz==2020.4
sqlparse==0.4.1
static3==0.7.0
fontawesomefree==6.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-extensions==3.2.0
channels==2.4.0
Werkzeug==2.0.3
django-weasyprint==2.2.0
dynaconf==3.1.12
ipython==8.5.0
django-simple-history==3.3.0
dpd-components==0.1.0
dash-bootstrap-components==0.13.1
plotly==5.13.1
dash==2.6.1
numpy==1.23.4
pandas==1.5.1
django-plotly-dash==2.2.0
django-redis==5.3.0
Unidecode==1.3.6